@@ -5,15 +5,14 @@
   "dependencies": {
     "@types/jest": "^27.5.2",
     "@types/node": "^17.0.45",
-    "@types/react-dom": "^18.2.7",
-    "babel-preset-react-app": "^10.0.1",
+    "@types/react-dom": "19.1.5",
+    "babel-preset-react-app": "10.1.0",
     "normalize.css": "^8.0.1",
-    "react": "18.3.0",
-    "react-dom": "18.3.0",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "react-scripts": "^5.0.1",
     "react-sticky": "^6.0.3",
-    "styled-components": "^6.1.18",
-    "typescript": "^4.9.5"
+    "styled-components": "^6.1.18"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -28,6 +27,7 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@types/react": "^18.2.20"
+    "@types/react": "19.1.6",
+    "typescript": "^4.9.5"
   }
 }
